Starting to work with two modules: extractr and locatr.
#####Extractr:

Right now, all this does, really, is open a connection to the Scala app and pull down some JSON. I've stubbed out a method for submitting queries into Scala as well, but that's code for another evening.

On the up side: this totally works. Huzzah! JSON is easily retreaved and read from Scala, and that's pretty neat. This module will _eventually_ form and GET query parameters against the Scala app.

#####Locatr:

This... is getting there, but in no means ready for prime time. This will, eventually, perform the task of trying to figure out if we Have a Thing, or if we need to make it. More specifically, it should integrate somehow with Model#find, so that on find_by_name we have this flow:
* Check DB
* rescue ActiveRecord::RecordNotFound
** Query against Scala
** Load result if one is returned
* THEN, if record is still not found, tell the user so.

This commit also adds create_from_hash and create_from_json methods to the Recipe model. Create_from_json does *not* require symbolizing hash keys.